Make RequestScopedPartialStreetEdge inherit from PartialStreetEdge
