+++ src/main/java/org/opentripplanner/routing/edgetype/RequestScopedPartialStreetEdge.java
@@ -1,28 +1,14 @@
 package org.opentripplanner.routing.edgetype;
 
-import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.LineString;
-import org.opentripplanner.common.TurnRestriction;
-import org.opentripplanner.routing.graph.Edge;
-import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.util.ElevationUtils;
 import org.opentripplanner.routing.vertextype.StreetVertex;
 import org.opentripplanner.routing.vertextype.RequestScopedVertex;
 import org.opentripplanner.util.I18NString;
 
-import java.util.List;
-
-
-final public class RequestScopedPartialStreetEdge extends StreetWithElevationEdge implements
+final public class RequestScopedPartialStreetEdge extends PartialStreetEdge implements
     RequestScopedEdge {
 
     private static final long serialVersionUID = 1L;
-
-    /**
-     * The edge on which this lies.
-     */
-    private StreetEdge parentEdge;
-
 
     /**
      * Create a new partial street edge along the given 'parentEdge' from 'v1' to 'v2'.
@@ -31,10 +17,7 @@ final public class RequestScopedPartialStreetEdge extends StreetWithElevationEdg
      */
     public RequestScopedPartialStreetEdge(StreetEdge parentEdge, StreetVertex v1, StreetVertex v2,
             LineString geometry, I18NString name, double length) {
-        super(v1, v2, geometry, name, length, parentEdge.getPermission(), false);
-        this.parentEdge = parentEdge;
-        setCarSpeed(parentEdge.getCarSpeed());
-        setElevationProfileUsingParents();
+        super(parentEdge, v1, v2, geometry, name, length);
 
         // Assert that the edge is going in the right direction [only possible if vertex is temporary]
         assertEdgeIsNotDirectedAwayFromTemporaryEndVertex(v1);
@@ -48,93 +31,18 @@ final public class RequestScopedPartialStreetEdge extends StreetWithElevationEdg
      */
     RequestScopedPartialStreetEdge(StreetEdge parentEdge, StreetVertex v1, StreetVertex v2,
             LineString geometry, I18NString name) {
-        super(v1, v2, geometry, name, 0, parentEdge.getPermission(), false);
-        this.parentEdge = parentEdge;
-        setCarSpeed(parentEdge.getCarSpeed());
-
-        // No length is known, so we use the provided geometry to estimate it
-        calculateLengthFromGeometry();
-        setElevationProfileUsingParents();
+        super(parentEdge, v1, v2, geometry, name);
 
         // Assert that the edge is going in the right direction [only possible if vertex is temporary]
         assertEdgeIsNotDirectedAwayFromTemporaryEndVertex(v1);
         assertEdgeIsDirectedTowardsTemporaryEndVertex(v2);
     }
 
-    /**
-     * Partial edges are always partial.
-     */
-    @Override
-    public boolean isPartial() {
-        return true;
-    }
-
-    /**
-     * Have the inbound angle of  their parent.
-     */
-    @Override
-    public int getInAngle() {
-        return parentEdge.getInAngle();
-    }
-
-    /**
-     * Have the outbound angle of  their parent.
-     */
-    @Override
-    public int getOutAngle() {
-        return parentEdge.getInAngle();
-    }
-
-    /**
-     * Have the turn restrictions of  their parent.
-     */
-    @Override
-    protected List<TurnRestriction> getTurnRestrictions(Graph graph) {
-        return graph.getTurnRestrictions(parentEdge);
-    }
-
-    /**
-     * This implementation makes it so that TurnRestrictions on the parent edge are applied to this edge as well.
-     */
-    @Override
-    public boolean isEquivalentTo(Edge e) {
-        return (e == this || e == parentEdge);
-    }
-
-    @Override
-    public boolean isReverseOf(Edge e) {
-        Edge other = e;
-        if (e instanceof RequestScopedPartialStreetEdge) {
-            other = ((RequestScopedPartialStreetEdge) e).parentEdge;
-        }
-
-        // TODO(flamholz): is there a case where a partial edge has a reverse of its own?
-        return parentEdge.isReverseOf(other);
-    }
-
-    @Override
-    public boolean isRoundabout() {
-        return parentEdge.isRoundabout();
-    }
-
-    /**
-     * Returns true if this edge is trivial - beginning and ending at the same point.
-     */
-    public boolean isTrivial() {
-        Coordinate fromCoord = this.getFromVertex().getCoordinate();
-        Coordinate toCoord = this.getToVertex().getCoordinate();
-        return fromCoord.equals(toCoord);
-    }
-
-    public StreetEdge getParentEdge() {
-        return parentEdge;
-    }
-
     @Override
     public String toString() {
-        return "TemporaryPartialStreetEdge(" + this.getName() + ", " + this.getFromVertex() + " -> "
+        return "RequestScopedPartialStreetEdge(" + this.getName() + ", " + this.getFromVertex() + " -> "
                 + this.getToVertex() + " length=" + this.getDistanceMeters() + " carSpeed="
-                + this.getCarSpeed() + " parentEdge=" + parentEdge + ")";
+                + this.getCarSpeed() + " parentEdge=" + super.getParentEdge() + ")";
     }
 
     private void assertEdgeIsNotDirectedAwayFromTemporaryEndVertex(StreetVertex v1) {
@@ -151,14 +59,5 @@ final public class RequestScopedPartialStreetEdge extends StreetWithElevationEdg
                 throw new IllegalStateException("A temporary edge is directed towards a start vertex");
             }
         }
-    }
-
-    private void setElevationProfileUsingParents() {
-        setElevationProfile(
-                ElevationUtils.getPartialElevationProfile(
-                        getParentEdge().getElevationProfile(), 0, getDistanceMeters()
-                ),
-                false
-        );
     }
 }
